requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
web3==6.11.1
solana==0.30.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-binance==1.0.19
aiohttp==3.9.1
websockets==12.0
ccxt==4.1.62
technical-analysis-tools==0.2.4
pynacl>=1.5.0
